+++ multichoose.cpp
@@ -1,4 +1,4 @@
 #include "choose.cpp"
-long int multichoose(int n, int k){
+unsigned long long int multichoose(int n, int k){
 	return choose(n+k-1, k);
 }
